+++ apps/block_scout_web/assets/js/lib/random_access_pagination.js
@@ -77,6 +77,8 @@ export function asyncReducer (state = asyncInitialState, action) {
         }
         delete action.nextPageParams.pageNumber
 
+        if (action.items.length === 0) hideLimitMessage()
+
         return Object.assign({}, state, {
           requestError: false,
           emptyResponse: action.items.length === 0,
@@ -87,6 +89,9 @@ export function asyncReducer (state = asyncInitialState, action) {
           beyondPageOne: pageNumber !== 1
         })
       }
+
+      if (action.items.length === 0) hideLimitMessage()
+
       return Object.assign({}, state, {
         requestError: false,
         emptyResponse: action.items.length === 0,
@@ -317,6 +322,10 @@ if ($element.length) {
     connectElements({ store, elements })
     firstPageLoad(store)
   }
+}
+
+function hideLimitMessage () {
+  $('[txs-limit]').hide()
 }
 
 function pagesNumbersGenerate (pagesLimit, $container, currentPageNumber, loading) {