+++ apps/block_scout_web/lib/block_scout_web/chain.ex
@@ -17,6 +17,8 @@ defmodule BlockScoutWeb.Chain do
 
   alias Explorer.Chain.Block.Reward
 
+  alias Explorer.{Chain, PagingOptions}
+
   alias Explorer.Chain.{
     Address,
     Address.CoinBalance,
@@ -30,8 +32,6 @@ defmodule BlockScoutWeb.Chain do
     Transaction,
     Wei
   }
-
-  alias Explorer.PagingOptions
 
   defimpl Poison.Encoder, for: Decimal do
     def encode(value, _opts) do
@@ -108,6 +108,23 @@ defmodule BlockScoutWeb.Chain do
       end
 
     Map.put(next_page_params, "items_count", items_count)
+  end
+
+  def next_page_params(params, transactions_count) do
+    if fetch_page_number(params) == 1 and !is_nil(transactions_count) do
+      page_size = Chain.default_page_size()
+
+      pages_limit = transactions_count |> Kernel./(page_size) |> Float.ceil() |> trunc()
+
+      %{}
+      |> Map.put("pages_limit", pages_limit)
+      |> Map.put("page_size", page_size)
+      |> Map.put("page_number", 1)
+    else
+      params
+      |> Map.delete("type")
+      |> Map.delete("address_id")
+    end
   end
 
   def paging_options(%{"hash" => hash, "fetched_coin_balance" => fetched_coin_balance}) do
@@ -234,6 +251,16 @@ defmodule BlockScoutWeb.Chain do
   end
 
   def fetch_page_number(_), do: 1
+
+  def supplement_page_options(options, params) do
+    options
+    |> Keyword.put(
+      :paging_options,
+      params
+      |> fetch_page_number()
+      |> update_page_parameters(Chain.default_page_size(), Keyword.get(options, :paging_options))
+    )
+  end
 
   def update_page_parameters(new_page_number, new_page_size, %PagingOptions{} = options) do
     %PagingOptions{options | page_number: new_page_number, page_size: new_page_size}