+++ apps/block_scout_web/lib/block_scout_web/controllers/address_transaction_controller.ex
@@ -5,10 +5,15 @@ defmodule BlockScoutWeb.AddressTransactionController do
 
   use BlockScoutWeb, :controller
 
-  import BlockScoutWeb.Chain, only: [current_filter: 1, paging_options: 1, next_page_params: 3, split_list_by_page: 1]
+  import BlockScoutWeb.Chain,
+    only: [
+      current_filter: 1,
+      next_page_params: 2,
+      supplement_page_options: 2
+    ]
 
   alias BlockScoutWeb.{AccessHelpers, Controller, TransactionView}
-  alias Explorer.{Chain, Market}
+  alias Explorer.{Chain, Market, PagingOptions}
 
   alias Explorer.Chain.{
     AddressInternalTransactionCsvExporter,
@@ -21,7 +26,8 @@ defmodule BlockScoutWeb.AddressTransactionController do
   alias Indexer.Fetcher.CoinBalanceOnDemand
   alias Phoenix.View
 
-  @transaction_necessity_by_association [
+  @default_options [
+    paging_options: %PagingOptions{page_size: Chain.default_page_size()},
     necessity_by_association: %{
       [created_contract_address: :names] => :optional,
       [from_address: :names] => :optional,
@@ -43,52 +49,28 @@ defmodule BlockScoutWeb.AddressTransactionController do
          {:ok, address} <- Chain.hash_to_address(address_hash, address_options, false),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
       options =
-        @transaction_necessity_by_association
-        |> Keyword.merge(paging_options(params))
+        @default_options
         |> Keyword.merge(current_filter(params))
+        |> supplement_page_options(params)
 
-      results_plus_one = Chain.address_to_transactions_with_rewards(address_hash, options)
-      {results, next_page} = split_list_by_page(results_plus_one)
+      %{transactions_count: transactions_count, transactions: transactions} =
+        Chain.address_to_transactions_rap(address_hash, options)
 
-      next_page_url =
-        case next_page_params(next_page, results, params) do
-          nil ->
-            nil
-
-          next_page_params ->
-            address_transaction_path(
-              conn,
-              :index,
-              address,
-              Map.delete(next_page_params, "type")
-            )
-        end
+      next_page_params = next_page_params(params, transactions_count)
 
       items_json =
-        Enum.map(results, fn result ->
-          case result do
-            {%Chain.Block.Reward{} = emission_reward, %Chain.Block.Reward{} = validator_reward} ->
-              View.render_to_string(
-                TransactionView,
-                "_emission_reward_tile.html",
-                current_address: address,
-                emission_funds: emission_reward,
-                validator: validator_reward
-              )
-
-            %Chain.Transaction{} = transaction ->
-              View.render_to_string(
-                TransactionView,
-                "_tile.html",
-                conn: conn,
-                current_address: address,
-                transaction: transaction,
-                burn_address_hash: @burn_address_hash
-              )
-          end
+        Enum.map(transactions, fn transaction ->
+          View.render_to_string(
+            TransactionView,
+            "_tile.html",
+            conn: conn,
+            current_address: address,
+            transaction: transaction,
+            burn_address_hash: @burn_address_hash
+          )
         end)
 
-      json(conn, %{items: items_json, next_page_path: next_page_url})
+      json(conn, %{items: items_json, next_page_params: next_page_params})
     else
       :error ->
         unprocessable_entity(conn)
@@ -99,7 +81,7 @@ defmodule BlockScoutWeb.AddressTransactionController do
       {:error, :not_found} ->
         case Chain.Hash.Address.validate(address_hash_string) do
           {:ok, _} ->
-            json(conn, %{items: [], next_page_path: ""})
+            json(conn, %{items: [], next_page_params: nil})
 
           _ ->
             not_found(conn)