+++ apps/block_scout_web/test/block_scout_web/features/viewing_chain_test.exs
@@ -33,6 +33,7 @@ defmodule BlockScoutWeb.ViewingChainTest do
 
   describe "viewing addresses" do
     test "search for address", %{session: session} do
+      Application.put_env(:block_scout_web, BlockScoutWeb.Chain, has_emission_funds: false)
       address = insert(:address)
 
       start_supervised!(AddressesCounter)