+++ apps/block_scout_web/test/block_scout_web/features/viewing_transactions_test.exs
@@ -83,7 +83,7 @@ defmodule BlockScoutWeb.ViewingTransactionsTest do
       |> refute_has(TransactionListPage.transaction(pending))
     end
 
-    test "viewing the pending tranasctions list", %{
+    test "viewing the pending transactions list", %{
       pending: pending,
       pending_contract: pending_contract,
       session: session
@@ -154,6 +154,8 @@ defmodule BlockScoutWeb.ViewingTransactionsTest do
       session: session,
       transaction: transaction
     } do
+      Application.put_env(:block_scout_web, BlockScoutWeb.Chain, has_emission_funds: false)
+
       session
       |> TransactionLogsPage.visit_page(transaction)
       |> TransactionLogsPage.click_address(lincoln)